requirements.txt
pylint
pytest
click
black
pytest-cov
databricks-sql-connector
python-dotenv
requests